+++ packages/toolkit/src/plugin/index.ts
@@ -9,12 +9,14 @@ import {
 	Prettify,
 } from '../_shared/_super'
 
-export type AnyPlugin = Plugin<Any, Any>
+export type AnyPlugin = Plugin<Any, Any, Any>
 export type Plugins = Record<string, AnyPlugin>
+
 export const createPlugin = <
 	To extends ClippiumData['flags'],
 	From extends ClippiumData['flags'],
->( plugin: Plugin<To, From> ): Plugin<To, From> => plugin
+	Docs extends ClippiumData['flags'],
+>( plugin: Plugin<To, From, Docs> ): Plugin<To, From, Docs> => plugin
 
 type ParsedFlags<F extends ClippiumData['flags']> = Pick<ParsedDataArgv<{ flags: F }>, 'flags'>
 type PluginUtils<F extends ClippiumData['flags']> = Prettify<Omit<Parameters<Clippium<ClippiumData>['fn']>[0], 'flags' | 'commands' | 'positionals'> & ParsedFlags<F>>
@@ -22,16 +24,32 @@ type PluginUtils<F extends ClippiumData['flags']> = Prettify<Omit<Parameters<Cli
 export type Plugin<
 	To extends ClippiumData['flags'],
 	From extends ClippiumData['flags'],
+	Docs extends ClippiumData['flags'],
 > = {
 	/**
 	 * Plugin description
 	 */
-	desc?   : string
+	desc? : string
+	/**
+	 * Docs function for plugin
+	 *
+	 */
+	docs?   : {
+		/**
+		 * Add custom flags for use in plugin command
+		 */
+		flags?    : Docs
+		/**
+		 * Add usage examples for the plugin command
+		 */
+		examples? : ClippiumData['examples']
+		fn        : ( utils: PluginUtils<Docs> & { input: ClippiumData } ) => Promise<string>
+	}
 	/**
 	 * Convert function for plugin
 	 *
 	 */
-	convert: {
+	convert?: {
 		toData?   : {
 			/**
 			 * Add custom flags for use in plugin command