+++ packages/preset/colored/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @clippium/preset-colored
 
+## 1.0.0
+
+### Minor Changes
+
+- Add more documentation, fix issues, add docs commands to toolkit etc
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @clippium/color@0.1.0
+  - clippium@0.1.0
+
 ## 0.0.4
 
 ### Patch Changes