+++ packages/utils/updater/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @clippium/updater
 
+## 0.1.0
+
+### Minor Changes
+
+- Add more documentation, fix issues, add docs commands to toolkit etc
+
 ## 0.0.4
 
 ### Patch Changes